fix: SearchObjectFactory builds Solr search objects
This ensure that the generated search objects have the necessary methods
and properties. In this case, we need search object to have a getRecords
method so that user lists with records that are not grouped works can be
displayed. This, in turn requires properties and methods that do exist
on SearchObject_SolrSearcher but are absent from
SearchObject_BaseSearcher. If it turns out that SearchObjectFactory does
need to sometimes generate non-Solr search objects, then additional
control flow may be needed to enable this.