fix: each record's data is processed
The earlier check ascertained that no searchOptions had been allocated
to allow a records' data to be processed. If searchOptions were found,
they were assigned to this record, meaning that every record in a list
was assigned the search options of the very first record to be
processed. This meant all records were incorrectly assigned the sourceId
of the first one to be processed.

The current check also looks at the queryString, containing the record's
sourceId, meaning that each distinct record will be processed
individually, and that the correctId will be assigned. This then allows
for the check verifying that the record found is the record on the list
to return true, and for the record to be listed and rendered correctly.