feat: Grouping_Item instances contain edition
so that a patron can see the edition of the records when looking through
the 'Where is it' modal on the search results page, adds an 'edition'
property to the Grouping_Item class, and update getRawItemDataFromDB()
on GroupedWorkDriver instances so that edition is fetched from the
database and included in a scopedItem's data.

This ultimately adds the 'edition' property to the $summary variable in
the getCopyDetails method on the GroupedWork_AJAX class, allowing it to
be assigned to $interface, thus making it available to copyDetails.tpl.