feat: "Show Editions in the Where is it popup" setting
Adds the "Show Editions in the Where is it popup" setting to Grouped
Work Display Settings.

This toggles the boolean value of the separateItemsByEditionInWhereIsIt
column added to grouped_work_display_settings. It will later allow the
user to set the "Where is it?" popup to be divided into sections by
editions, so that edition information and covers can be displayed.

To test, choose a Grouped Work Display Setting profile to edit (eg.
public). First check that its separateItemsByEditionInWhereIsIt value is
set to 0 (intended default) in the db. Then navigate  to Grouped Work
Display Settings > <chosen setting profile> > Searching > Search
Results, and ensure that the "Show Editions in the Where is it popup"
checkbox displays and is toggled off. Hover over the interrogation mark
to get an explanation of the expected behaviour change. Toggle on the
setting, and save your changes. In the db, check that the value of
separateItemsByEditionInWhereIsIt for the setting profile you edited is
now set to 1.