feat: display copies per edition per location in Where is it popup
In search results, for the list of records grouped into a specific
title, display each of them in a separate section of the Where is it
popup. Each section includes a header containing edition information,
then lists the available copies of this editions, grouped by the
location they are found at.

This ultimately allows edition information to be displayed within the
Where is it popup (as per the enhancement request).

Test plan:

In Aspen Administration > Grouped Work Display Settings > Search Results
- have "Show Editions in the Where is it popup" toggled on
- have "Copy Information to show" set to something that allows for
"Where is it" links to display
- (if you'd like for covers to be included, have "Show Covers for
Editions" toggled on)

Run a search in Library Catalog, and open "Where is it?" links on
various titles. Notice that the table displayed in the popup is now
subvided in 'edition' sections. These sections contain a list of copies
grouped by location and their summary information. (Toggled off "Show
Edditions in the Where is it popup" to see the original behaviour and
table structure.)

Note: this does not affect "Where is it?" popups opened for individual
records.