feat: display edition info for individual records in Where is it popup
The 'Where is it?' link also appear for individual records, a list of
which can be displayed for a specific title by selecting the 'Show
Edition(s)' button. This applies the changes in 'feat: display copies
per edition per location in Where is it popup' to 'Where is it?' popups
opened through links on individual records, allowing for edition
information to be shown.

Test plan: same as for 'feat: display copies per edition per location in
Where is it popup', but focusing on 'Where is it' links on individual
records.