feat: unique username field id
A user might attempt to sign in while viewing the PIN email reset page
(MyAccount/EmailResetPin). Without this patch, any attempt at submitting
login info through the login form popup while on said page will fail.

This is because there is both a username field in the PIN email reset
form and in the login form, and both share the same id. Since the log
in form is referred to as the 'ajax-login' template, this patch
changes the id of its username field to 'ajaxLoginUsername', so that
it is unique in this scenario.

To replicate:
- go to /MyAccount/EmailResetPin
- select 'Sign In'
- fill in the form with valid credentials
- notice the 'Please enter both Library Barcode and PIN / Password.'
  error message.

Test plan:

- apply this patch
- repeat the steps above
- notice that, instead of an error, you are logged in as normal.